+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.ReservationTimeCreateRequest;
-import roomescape.dto.ReservationTimeResponse;
+import roomescape.requestDto.ReservationTimeCreateRequest;
+import roomescape.responseDto.ReservationTimeResponse;
 import roomescape.service.ReservationTimeService;
 
 @RestController
