+++ src/main/java/roomescape/controller/ThemeController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.ThemeCreateRequest;
-import roomescape.dto.ThemeResponse;
+import roomescape.requestDto.ThemeCreateRequest;
+import roomescape.responseDto.ThemeResponse;
 import roomescape.service.ThemeService;
 
 @RestController
