+++ src/main/java/roomescape/requestDto/ThemeCreateRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.requestDto;
 
 public record ThemeCreateRequest(String name, String description, String thumbnail) {
 }
