+++ src/main/java/roomescape/responseDto/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.responseDto;
 
 import java.time.format.DateTimeFormatter;
 import roomescape.domain.ReservationTime;
