+++ src/main/java/roomescape/responseDto/ThemeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.responseDto;
 
 import roomescape.domain.Theme;
