+++ src/main/java/roomescape/service/ReservationService.java
@@ -10,9 +10,9 @@ import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
 import roomescape.domain.UserName;
-import roomescape.dto.AvailableTimeResponse;
-import roomescape.dto.ReservationCreateRequest;
-import roomescape.dto.ReservationResponse;
+import roomescape.responseDto.AvailableTimeResponse;
+import roomescape.requestDto.ReservationCreateRequest;
+import roomescape.responseDto.ReservationResponse;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
