+++ src/main/java/roomescape/service/ThemeService.java
@@ -3,8 +3,8 @@ package roomescape.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.domain.Theme;
-import roomescape.dto.ThemeCreateRequest;
-import roomescape.dto.ThemeResponse;
+import roomescape.requestDto.ThemeCreateRequest;
+import roomescape.responseDto.ThemeResponse;
 import roomescape.repository.ThemeRepository;
 
 @Service
