+++ src/main/resources/data.sql
@@ -20,34 +20,34 @@ INSERT INTO reservation_time (start_at) VALUES ('15:00');
 INSERT INTO reservation_time (start_at) VALUES ('17:00');
 INSERT INTO reservation_time (start_at) VALUES ('18:00');
 
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', '2024-04-26', '1', '1');
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저02', '2024-04-26', '2', '1');
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저03', '2024-04-26', '3', '1');
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저04', '2024-04-26', '4', '1');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', DATEADD(DAY, -7, CURDATE()), '1', '1');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저02', DATEADD(DAY, -7, CURDATE()), '2', '1');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저03', DATEADD(DAY, -7, CURDATE()), '3', '1');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저04', DATEADD(DAY, -7, CURDATE()), '4', '1');
 
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저03', '2024-04-26', '3', '7');
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저03', '2024-04-26', '3', '8');
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저03', '2024-04-26', '3', '9');
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저03', '2024-04-26', '3', '10');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저03', DATEADD(DAY, -7, CURDATE()), '3', '7');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저03', DATEADD(DAY, -7, CURDATE()), '3', '8');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저03', DATEADD(DAY, -7, CURDATE()), '3', '9');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저03', DATEADD(DAY, -7, CURDATE()), '3', '10');
 
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', '2024-04-27', '1', '2');
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저02', '2024-04-27', '2', '2');
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저03', '2024-04-27', '3', '2');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', DATEADD(DAY, -6, CURDATE()), '1', '2');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저02', DATEADD(DAY, -6, CURDATE()), '2', '2');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저03', DATEADD(DAY, -6, CURDATE()), '3', '2');
 
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', '2024-04-28', '1', '3');
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저02', '2024-04-28', '2', '3');
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저03', '2024-04-28', '3', '3');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', DATEADD(DAY, -5, CURDATE()), '1', '3');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저02', DATEADD(DAY, -5, CURDATE()), '2', '3');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저03', DATEADD(DAY, -5, CURDATE()), '3', '3');
 
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', '2024-04-29', '1', '4');
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저02', '2024-04-29', '2', '4');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', DATEADD(DAY, -4, CURDATE()), '1', '4');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저02', DATEADD(DAY, -4, CURDATE()), '2', '4');
 
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', '2024-04-30', '1', '5');
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저02', '2024-04-30', '2', '5');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', DATEADD(DAY, -3, CURDATE()), '1', '5');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저02', DATEADD(DAY, -3, CURDATE()), '2', '5');
 
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', '2024-05-01', '1', '6');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', DATEADD(DAY, -2, CURDATE()), '1', '6');
 
 INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', '2024-04-01', '1', '15');
-INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', '2024-05-05', '1', '11');
+INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', DATEADD(DAY, -1, CURDATE()), '1', '11');
 
 INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저01', '2023-04-26', '1', '1');
 INSERT INTO reservation (name, date, time_id, theme_id) VALUES ('유저02', '2023-04-26', '2', '1');
